feat!: delete helpers.py, move functions to their natural homes
Eliminate the catch-all helpers module by moving each function to the
module that actually uses it:

- utc_now() → inlined as datetime.now(timezone.utc) at all call sites
- wait_for_notice_event() → pgqueuer.core.listeners
- timeout_with_jitter() → inlined into pgqueuer.core.buffers
- normalize_cron_expression() → inlined into pgqueuer.core.sm
- add_schema_to_dsn() → pgqueuer.adapters.cli.cli
- merge_tracing_headers() → pgqueuer.adapters.persistence.query_helpers
- timer(), ExponentialBackoff, retry_timer_buffer_timeout() → deleted
  (dead code, only used in tests for removed features)
